refactor(client/component): Re-order placement of some component
Because some file are in shared, but their aren't used in both part.
