+++ Engine/Client/CMakeLists.txt
@@ -5,9 +5,39 @@ set(EXE r-type_client)
 set(EXE_SRC_PATH "Src")
 set(EXE_INCLUDE_PATH "Include")
 
-set(COMPONENT_SRC_FILE ./Src/Components/ShaderComponent/ShaderPathComponent.cpp)
+set(COMPONENT_SRC_FILE  ./Src/Components/ShaderComponent/ShaderPathComponent.cpp
+                        ./Src/Components/Size1DComponent/Size1DComponent.cpp
+                        ./Src/Components/TextComponent/TextComponent.cpp
+                        ./Src/Components/MaterialMapComponent/MaterialMapComponent.cpp
+                        ./Src/Components/MusicPathComponent/MusicPathComponent.cpp
+                        ./Src/Components/MusicPitchComponent/MusicPitchComponent.cpp
+                        ./Src/Components/MusicVolumeComponent/MusicVolumeComponent.cpp
+                        ./Src/Components/SoundPathComponent/SoundPathComponent.cpp
+                        ./Src/Components/SoundPitchComponent/SoundPitchComponent.cpp
+                        ./Src/Components/SoundVolumeComponent/SoundVolumeComponent.cpp
+                        ./Src/Components/FontPathComponent/FontPathComponent.cpp
+                        ./Src/Components/ObjPathComponent/ObjPathComponent.cpp
+                        ./Src/Components/TexturePathComponent/TexturePathComponent.cpp
+                        ./Src/Components/ScaleComponent/ScaleComponent.cpp
+                        ./Src/Components/ColourComponent/ColourComponent.cpp
+)
 
-set(COMPONENT_INCLUDE_DIR ./Include/Components/ShaderComponent)
+set(COMPONENT_INCLUDE_DIR   ./Src/Components/ShaderComponent
+                            ./Src/Components/Size1DComponent/
+                            ./Src/Components/TextComponent/
+                            ./Src/Components/MaterialMapComponent/
+                            ./Src/Components/MusicPathComponent/
+                            ./Src/Components/MusicPitchComponent/
+                            ./Src/Components/MusicVolumeComponent/
+                            ./Src/Components/SoundPathComponent/
+                            ./Src/Components/SoundPitchComponent/
+                            ./Src/Components/SoundVolumeComponent/
+                            ./Src/Components/FontPathComponent/
+                            ./Src/Components/ObjPathComponent/
+                            ./Src/Components/TexturePathComponent/
+                            ./Src/Components/ScaleComponent/
+                            ./Src/Components/ColourComponent/
+)
 
 set(CMAKE_CXX_STANDARD 20)
 
@@ -17,22 +47,8 @@ set(SRC_EXE
         Src/ClientSceneManager/ClientSceneManager.cpp
         Src/GraphicalLibrary/GetGraphicalLibrary.cpp
         ../Shared/SceneManager/ASceneManager.cpp
-        ../Shared/Component/FontPathComponent.cpp
-        ../Shared/Component/ObjPathComponent.cpp
-        ../Shared/Component/TexturePathComponent.cpp
-        ../Shared/Component/ScaleComponent.cpp
-        ../Shared/Component/ColourComponent.cpp
         ../Shared/Component/Position3DComponent.cpp
         ../Shared/Component/Position2DComponent.cpp
-        ../Shared/Component/Size1DComponent.cpp
-        ../Shared/Component/TextComponent.cpp
-        ../Shared/Component/MaterialMapComponent.cpp
-        ../Shared/Component/MusicPathComponent.cpp
-        ../Shared/Component/MusicPitchComponent.cpp
-        ../Shared/Component/MusicVolumeComponent.cpp
-        ../Shared/Component/SoundPathComponent.cpp
-        ../Shared/Component/SoundPitchComponent.cpp
-        ../Shared/Component/SoundVolumeComponent.cpp
         ${COMPONENT_SRC_FILE}
 )
