+++ Engine/Client/Src/Components/FontPathComponent/FontPathComponent.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "IComponent.hpp"
+#include "../../../../Shared/Interface/IComponent.hpp"
 
 /**
  * @class FontPathComponent
