+++ Engine/Client/Src/Components/MusicPathComponent/MusicPathComponent.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "IComponent.hpp"
+#include "../../../../Shared/Interface/IComponent.hpp"
 
 /**
  * @class MusicPathComponent
