+++ Engine/Client/Src/Components/MusicPitchComponent/MusicPitchComponent.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "IComponent.hpp"
+#include "../../../../Shared/Interface/IComponent.hpp"
 
 /**
  * @class MusicPitchComponent
