+++ Engine/Client/Src/Components/MusicVolumeComponent/MusicVolumeComponent.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "IComponent.hpp"
+#include "../../../../Shared/Interface/IComponent.hpp"
 
 /**
  * @class MusicVolumeComponent
