+++ Engine/Client/Src/Components/ScaleComponent/ScaleComponent.cpp
