+++ Engine/Client/Src/Components/Size1DComponent/Size1DComponent.cpp
