+++ Engine/Client/Src/Components/Size1DComponent/Size1DComponent.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "IComponent.hpp"
+#include "../../../../Shared/Interface/IComponent.hpp"
 
 /**
  * @class Size1DComponent
