+++ Engine/Client/Src/Components/SoundVolumeComponent/SoundVolumeComponent.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "IComponent.hpp"
+#include "../../../../Shared/Interface/IComponent.hpp"
 
 /**
  * @class SoundVolumeComponent
