+++ Engine/Shared/SceneManager/ASceneManager.hpp
@@ -13,22 +13,22 @@
 #include "ISceneManager.hpp"
 #include "../Interface/ISystem.hpp"
 #include "../Enum/StringKeyMap.hpp"
-#include "TextComponent.hpp"
-#include "ScaleComponent.hpp"
-#include "ColourComponent.hpp"
-#include "Size1DComponent.hpp"
-#include "ObjPathComponent.hpp"
-#include "FontPathComponent.hpp"
-#include "MusicPathComponent.hpp"
-#include "SoundPathComponent.hpp"
-#include "MusicPitchComponent.hpp"
+#include "../../Client/Src/Components/TextComponent/TextComponent.hpp"
+#include "../../Client/Src/Components/ScaleComponent/ScaleComponent.hpp"
+#include "../../Client/Src/Components/ColourComponent/ColourComponent.hpp"
+#include "../../Client/Src/Components/Size1DComponent/Size1DComponent.hpp"
+#include "../../Client/Src/Components/ObjPathComponent/ObjPathComponent.hpp"
+#include "../../Client/Src/Components/FontPathComponent/FontPathComponent.hpp"
+#include "../../Client/Src/Components/MusicPathComponent/MusicPathComponent.hpp"
+#include "../../Client/Src/Components/SoundPathComponent/SoundPathComponent.hpp"
+#include "../../Client/Src/Components/MusicPitchComponent/MusicPitchComponent.hpp"
 #include "Position2DComponent.hpp"
 #include "Position3DComponent.hpp"
-#include "SoundPitchComponent.hpp"
-#include "MaterialMapComponent.hpp"
-#include "MusicVolumeComponent.hpp"
-#include "SoundVolumeComponent.hpp"
-#include "TexturePathComponent.hpp"
+#include "../../Client/Src/Components/SoundPitchComponent/SoundPitchComponent.hpp"
+#include "../../Client/Src/Components/MaterialMapComponent/MaterialMapComponent.hpp"
+#include "../../Client/Src/Components/MusicVolumeComponent/MusicVolumeComponent.hpp"
+#include "../../Client/Src/Components/SoundVolumeComponent/SoundVolumeComponent.hpp"
+#include "../../Client/Src/Components/TexturePathComponent/TexturePathComponent.hpp"
 
 #define CONFIG_SUFFIX ".json"
